wrap API calls with location check, add some comments in env-sample

+++ team_drive.rb
@@ -13,7 +13,7 @@ team_info["members"].each do |entry|
 end
 
 # prepare for authentication
-user_id = ENV['YOUR_ID']
+user_id = ENV['USERNAME']
 
 OOB_URI = 'urn:ietf:wg:oauth:2.0:oob'
 scope = 'https://www.googleapis.com/auth/drive'
@@ -42,17 +42,23 @@ drive_service.authorization = credentials
 # list of permission for this team drive
 permission_list = drive_service.list_permissions(ENV["DRIVE_ID"], supports_team_drives: true)
 
+puts "Deleting emails no longer in team.yml:"
 permission_list.items.each do |permission|
   this_email = permission.email_address
   # if email is not in team's emails set, remove permission
   unless team_emails.include?(this_email)
-    drive_service.delete_permission(ENV["DRIVE_ID"], permission.id, supports_team_drives: true)
+    puts "deleting #{this_email}"
+    if ENV['LOCATION'] == 'PRODUCTION'
+      puts "Making API call..."
+      drive_service.delete_permission(ENV["DRIVE_ID"], permission.id, supports_team_drives: true)
+    end
   else
     team_emails.delete(this_email)
   end
 end
 
 # if there's any email left in the set, create new permission
+puts "Adding new emails"
 unless team_emails.empty?
   team_emails.each do |email|
     permission_detail = {
@@ -61,12 +67,15 @@ unless team_emails.empty?
       type: "user",
       value: email
     }
-    new_permission = Google::Apis::DriveV2::Permission.new(permission_detail)
-    drive_service.insert_permission(ENV["DRIVE_ID"], new_permission, supports_team_drives: true, send_notification_emails: true) do |result, err|
-      if err
-        puts "Error: #{err}"
-      else
-        puts "Successfully adding #{result.name} with email #{result.email_address}"
+    puts "Adding #{email}"
+    if ENV['LOCATION'] == 'PRODUCTION'
+      new_permission = Google::Apis::DriveV2::Permission.new(permission_detail)
+      drive_service.insert_permission(ENV["DRIVE_ID"], new_permission, supports_team_drives: true, send_notification_emails: true) do |result, err|
+        if err
+          puts "Error: #{err}"
+        else
+          puts "Successfully made API call to add #{result.name} with email #{result.email_address}"
+        end
       end
     end
   end
